clear() and setImgSrc() cleanup
Because clear() is bypassed when draw(true) is called, the
olImageLoadError class is better removed in setImgSrc(). Same goes for the
alphaHack filter. It also turned out that both setImgSrc() and clear()
remove the image from the dom, so this can be removed from clear()
because clear() also calls setImgSrc().