Simplified logic for detecting cached tiles on a backbuffer
We no longer need to bail out when the img parent is the current
backbuffer. That was added to fix #980, but the real fix for #980 is to not
manage the tile cache in addLayer when a tile has an imgDiv already. In
addition, it is not needed to manage the tile cache in removeLayer.