Use merged temp file for compilers that operate on files.
The closure and uglify-js compilers need a list of files, not an
in-memory string.  The option --amd=pre tries to wrap the merged
contents with AMD boilerplate, but if either of these compilers was
selected there is no merged contents.  Instead, always merge files,
and just write it to a temp file for any compiler that needs it.

See #1033.