Adding refresh event and making TileManager listen for it
When calling redraw(true) on a HTTPRequest layer, mergeNewParams is used to
add a url parameter with a random value. For layer types that do not use
params in the getURL function, the url does not change and the TileManager
will provide tiles from the cache. To make sure that tiles are fetched from
the server, they need to be removed from the cache first.