Correct handling of 0 and Infinity in Min/MaxScaleDenominator
minScale and maxScale values of 0 and Infinity do not make sense in
OpenLayers, but they are used in WMS GetCapabilities 1.3.0. to indicate
that there is no minScale or maxScale restriction. Assuming that the
capability values for minScale and maxScale are passed as config options
to a WMS layer, it is better to not set minScale and maxScale when 0 is
found in MinScaleDenominator or Infinity is found in MaxScaleDenominator.