Do not draw destroyed tiles
Too keep track of destroyed tiles, the TileManager registers for the
unload event of each tile. However, that is only fired when the tile is
not loaded yet upon destruction. By checking for tile.events, we know when
a tile has been destroyed, and we can remove it from the queue without
attempting to draw it.