zoomStart event and close-to-synced zoom for EventPane layers
With a zoomStart event we can notify EventPane layers like Google of
a starting zoom. This gives the Google layer a chance to start its built-in
zoom animation at the same time that the OpenLayers map starts its
animiation, resulting in a better sync between Google and OpenLayers zoom
animation.