Add cluster handling to filter
This filters the items in a cluster.  I wrote this because handleAdd
gets rerun when using filters before clusters.  Not having this
handling was causing the filter to fail (or succeed, depending on your
filter) on each clustered feature since it was inspecting the cluster
and not the features within the cluster.