Promote several properties to API properties.
The properties `multipleKey` and `toggleKey` are used inside the examples
and therefore publicly known anyhow. The properties `onBeforeSelect` and
`scope` are changed to be part of the API for consistency (e.g. `onUnselect`
is part of the API already) and usefullness (we allow the `scope` to be
defined elsewhere in the library as well). The property `renderIntent` is
being promoted as it is very useful and a nice complimentary of the already
public `selectStyle`.