support for projectionless maps was broken
We had to monkey-patch Layer.setMap(), because we were using a projectionless layer {projection:null, units: 'm'}, where the map only knows the units of the layer. According to the docs, this should work, but we were getting a JS error in .setMap()/Line 639, because OpenLayers tries to extract the units from the projection object.