test GetFeatureInfo with invalid HTML for IE11
I think I experience the same problem of issue #1177

As an example I modified the GetFeatureInfo example injecting the corrupt data that cause the error in IE11. Just click in the map.

Obviously the not valid HTML5 is can't be read with DomParser() in IE11.
One solution is pull request #1285

Another solution may be to check the INFO_FORMAT and do not call
var features = this.format.read(doc);
if it is text/html or text/plain (i.e. not a GML) as in showInfo() just the text is used (no GML features):
document.getElementById('responseText').innerHTML = evt.text;