Added imagery metadata at center map functionality
Added metadataCenter property to OpenLayers.Layer.Bing
Updated attributionTemplate to fit vintageStart and vintageEnd dates. This can be further tuned.
Updated loadMetadata to use the OpenLayers.Util.createScript(which basically includes what was previously here).
Added loadMetadataCenter function which loads the metadata for specific point.
Updated updateAttribution so it also take care of the dates. I think it makes sense like this, but a new function can be created in the future.
Updated processMetadata so it fits both loadMetadata and loadMetadataCenter