Update Spherical.js
The current implementation of OpenLayers.Spherical.computeHeading() is wrong based on the relevant [link](http://www.movable-type.co.uk/scripts/latlong.html) listed in Spherical.js. The delta of the longitudes is reversed in both the 'y' and the 'x' portion of the equations. Instead of `from.lon - to.lon`, it should be `to.lon - from.lon`.