Only traverse ancestors so long as they are elements
In Firefox an error "does not implement interface Element" was
being raised when the MouseWheel event was outside the map, due
to the traverse up the tree winding up at the document element.

This error was caught in the try-catch block and so did not
bubble up to the user, but trying to use Firebug and stop on
all errors then it caused some irritation.