Fix pointer event handling for IE11
IE11 (and potentially soon other browsers) implement the pointer event
model using according to the W3C specification and not the MS-prefixed
events as implemented in IE10.

This commit fixes the mapping of browser pointer* events relating to
touch (down, move, up) to the touch* events (start, move, end) as used
in OpenLayers. The differentiation of pointer events with pointer type
touch is also corrected to reflect the W3C specification and current
usage.