Correct drag handling when in multi-touch pointer-model browsers
In the touch model, a multi-touch event is modelled as a single event
whereas in the pointer model each pointer touch is modelled as a
separate event with a distinct pointer ID. This led to problems as the
first touch of a multi-touch event is not recognised as being part of a
multi-touch event and therefore could lead to a mixture e.g.
drag/pinch-zoom.

This commit modifies the drag handler such that the pointerId of touch
events will be recorded for the first touch and subsequent touches with
a different pointerId (e.g. other fingers in a multi-touch event) can be
ignored.